mm: Change type of thp_size to size_t
ARM defines size_t to be unsigned int, which is exactly as many bits as
unsigned long, but causes a type mismatch when used with the min() macro.
Since it's a byte count, thp_size should be size_t.

+++ include/linux/huge_mm.h
@@ -474,7 +474,7 @@ static inline bool thp_migration_supported(void)
  *
  * Return: Number of bytes in this page.
  */
-static inline unsigned long thp_size(struct page *page)
+static inline size_t thp_size(struct page *page)
 {
 	return PAGE_SIZE << thp_order(page);
 }
